requirements.txt
bibt-gcp-bq==0.0.24
bibt-gcp-pubsub==0.0.19
bibt-gcp-secrets==0.0.12
bibt-gcp-storage==0.0.12
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
Flask==3.0.3
functions-framework==3.7.0
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-appengine-logging==1.4.3
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-cloud-functions==1.16.0
google-cloud-logging==3.10.0
google-cloud-pubsub==2.19.0
google-cloud-secret-manager==2.17.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.8.2
requests==2.32.3
rsa==4.9
six==1.16.0
urllib3==2.2.1
watchdog==4.0.1
Werkzeug==3.0.3
xmltodict==0.13.0
